Delete zombies only if there are more than 10000

+++ lnnode/zombies.go
@@ -5,6 +5,10 @@ import (
 	"github.com/lightningnetwork/lnd/channeldb"
 )
 
+const (
+	maxZombies = 10000
+)
+
 var (
 	edgeBucket   = []byte("graph-edge")
 	zombieBucket = []byte("zombie-index")
@@ -16,7 +20,14 @@ func deleteZombies(chanDB *channeldb.DB) error {
 		if edges == nil {
 			return channeldb.ErrGraphNoEdgesFound
 		}
-		return edges.DeleteBucket(zombieBucket)
+		zombies := edges.Bucket(zombieBucket)
+		if zombies == nil {
+			return nil
+		}
+		if zombies.Stats().KeyN > maxZombies {
+			return edges.DeleteBucket(zombieBucket)
+		}
+		return nil
 	})
 	return err
 }
